# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AVITO API",
        "description": "API endpoints for AVITO",
        "contact": {
            "email": "pashaboga4ev@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://0.0.0.0:80/api",
            "description": "Base server for development"
        }
    ],
    "paths": {
...
        "/tag": {
            "get": {
                "tags": [
                    "Avito Tag"
                ],
                "summary": "Get list of tags",
                "operationId": "tagsList",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Токен админа",
                        "schema": {
                            "type": "string",
                            "example": "admin_token"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/TagCollectionResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    }
                }
            },
            "post": {
                "tags": [
                    "Avito Tag"
                ],
                "summary": "Create tag",
                "operationId": "createTag",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Токен админа",
                        "schema": {
                            "type": "string",
                            "example": "admin_token"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/CreateTagRequestBody"
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/TagResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "422": {
                        "$ref": "#/components/responses/InvalidDataResponse"
                    }
                }
            }
        },
        "/tag/{banner_id}/attach-tags": {
            "post": {
                "tags": [
                    "Avito Tag"
                ],
                "summary": "Attach tags to banner",
                "operationId": "attachTagsToBanner",
                "parameters": [
                    {
                        "name": "banner_id",
                        "in": "path",
                        "description": "Banner ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Токен админа",
                        "schema": {
                            "type": "string",
                            "example": "admin_token"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/AttachDetachBannerTagsRequestBody"
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/TagWithoutPaginationCollectionResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    },
                    "422": {
                        "$ref": "#/components/responses/InvalidDataResponse"
                    }
                }
            }
        },
        "/tag/{id}": {
            "get": {
                "tags": [
                    "Avito Tag"
                ],
                "summary": "Get tag by id",
                "operationId": "getTag",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Токен админа",
                        "schema": {
                            "type": "string",
                            "example": "admin_token"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/TagResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    }
                }
            },
            "put": {
                "tags": [
                    "Avito Tag"
                ],
                "summary": "Update tag",
                "operationId": "updateTag",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Токен админа",
                        "schema": {
                            "type": "string",
                            "example": "admin_token"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateTagRequestBody"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/TagResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    },
                    "422": {
                        "$ref": "#/components/responses/InvalidDataResponse"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Avito Tag"
                ],
                "summary": "Delete tag",
                "operationId": "deleteTag",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Токен админа",
                        "schema": {
                            "type": "string",
                            "example": "admin_token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/SuccessDeletedResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    }
                }
            }
        },
        "/tag/{banner_id}/detach-tags": {
            "delete": {
                "tags": [
                    "Avito Tag"
                ],
                "summary": "Detach tags from banner",
                "operationId": "detachTags",
                "parameters": [
                    {
                        "name": "banner_id",
                        "in": "path",
                        "description": "Banner ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Токен админа",
                        "schema": {
                            "type": "string",
                            "example": "admin_token"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/AttachDetachBannerTagsRequestBody"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/TagWithoutPaginationCollectionResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "AttachDetachBannerTagsRequest": {
                "required": [
                    "tags"
                ],
                "properties": {
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": [
                            1,
                            2,
                            3
                        ]
                    }
                },
                "type": "object"
            },
...
            "CreateTagRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "tag name"
                    }
                },
                "type": "object"
            },
            "UpdateTagRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "tag name"
                    }
                },
                "type": "object"
            },
...
            "ResourceCollection": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "number",
                                "example": 2
                            },
                            "from": {
                                "type": "number",
                                "example": 1
                            },
                            "last_page": {
                                "type": "number",
                                "example": 8
                            },
                            "links": {
                                "type": "object"
                            },
                            "path": {
                                "type": "string",
                                "example": "http://0.0.0.0/api/v1/object/list?page=2"
                            },
                            "per_page": {
                                "type": "number",
                                "example": 1
                            },
                            "to": {
                                "type": "number",
                                "example": 1
                            },
                            "total": {
                                "type": "number",
                                "example": 1
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "InvalidData": {
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Any error message"
                    }
                },
                "type": "object"
            },
            "NotFound": {
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Not Found"
                    }
                },
                "type": "object"
            },
...
            "Unauthorized": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    }
                },
                "type": "object"
            },
            "TagCollection": {
                "required": [
                    "data"
                ],
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ResourceCollection"
                    },
                    {
                        "properties": {
                            "data": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/TagResource"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "TagResource": {
                "required": [
                    "tag_id",
                    "name",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "tag_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "tag name"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "TagWithoutPaginationCollection": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TagResource"
                        }
                    }
                },
                "type": "object"
            }
        },
        "responses": {
...
            "InvalidDataResponse": {
                "description": "Некорректные данные",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/InvalidData"
                        }
                    }
                }
            },
            "NotFoundResponse": {
                "description": "Not Found",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/NotFound"
                        }
                    }
                }
            },
...
            "SuccessDeletedResponse": {
                "description": "Success Deleted"
            },
            "UnauthorizedResponse": {
                "description": "Пользователь не авторизован",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Unauthorized"
                        }
                    }
                }
            },
            "TagCollectionResponse": {
                "description": "Success",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/TagCollection"
                        }
                    }
                }
            },
            "TagResponse": {
                "description": "Success",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/TagResource"
                        }
                    }
                }
            },
            "TagWithoutPaginationCollectionResponse": {
                "description": "Success",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/TagWithoutPaginationCollection"
                        }
                    }
                }
            }
        },
        "requestBodies": {
...
            "AttachDetachBannerTagsRequestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/AttachDetachBannerTagsRequest"
                        }
                    }
                }
            },
...
            "CreateTagRequestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/CreateTagRequest"
                        }
                    }
                }
            },
            "UpdateTagRequestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/UpdateTagRequest"
                        }
                    }
                }
            }
        }
    },
...
}